+++ .github/workflows/lint.yml
@@ -24,9 +24,9 @@ jobs:
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
       - name: Setup Go
-      - uses: actions/setup-go@v2
-      with:
-        go-version: 1.16.3
+        uses: actions/setup-go@v2
+        with:
+          go-version: 1.16.3
       - name: Checkout
         uses: actions/checkout@v2
       - name: Cache Go modules
