+++ .golangci.yml
@@ -103,14 +103,11 @@ linters-settings:
     line-length: 140
   maligned:
     suggest-new: true
-  misspell:
-    locale: US
   nolintlint:
     allow-leading-space: true # don't require machine-readable nolint directives (i.e. with no leading space)
     allow-unused: false # report any unused nolint directives
     require-explanation: false # don't require an explanation for nolint directives
     require-specific: false # don't require nolint directives to be specific about which linter is being skipped
-
 
 
 issues:
