+++ defaults.example.env
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+IRIS_MONGO_ADDR="mongo shard separated by ,"
+IRIS_MONGO_USER="mongo users"
+IRIS_MONGO_PASS="mongo password"
+IRIS_MONGO_DBNAME="mongo database name"
+
+IRIS_AUTHTOKEN="bot authtoken"
+IRIS_CLIENTID="bot clientid"
