--- hermodactylus/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "hermodactylus"
-version = "0.1.0"
-authors = ["Aaron Pham <aaronpham0103@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-[[bin]]
-name = "hermodactylus"
-path = "main.rs"
