--- pkg/pardanthopsis/pardanthopsis.go
@@ -1,1 +1,0 @@
-package pardanthopsis
