+++ web/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "web"
+version = "0.1.0"
+authors = ["Aaron Pham <aaronpham0103@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[[bin]]
+name = "web"
+path = "main.rs"
